content/smil/test/test_smilRestart.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <title>Test for SMIL Restart Behavior </title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
<svg id="svg" xmlns="http://www.w3.org/2000/svg" width="120px" height="120px">
  <!-- These 3 circles only differ in their animation's "restart" value -->
  <circle cx="20" cy="20" r="15" fill="blue">
    <animate attributeName="cx" from="20" to="100" begin="1s" dur="4s"
             restart="always" id="always" attributeType="XML"/>
  </circle>
  <circle cx="20" cy="60" r="15" fill="blue">
    <animate attributeName="cx" from="20" to="100" begin="1s" dur="4s"
             restart="whenNotActive" id="whenNotActive" attributeType="XML"/>
  </circle>
  <circle cx="20" cy="100" r="15" fill="blue">
    <animate attributeName="cx" from="20" to="100" begin="1s" dur="4s"
             restart="never" id="never" attributeType="XML"/>
  </circle>
</svg>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[
/** Test for SMIL Restart Behavior  **/

/* Global Variables */
var svg = document.getElementById("svg");
var always = document.getElementById("always");
var whenNotActive = document.getElementById("whenNotActive");
var never = document.getElementById("never");

SimpleTest.waitForExplicitFinish();

// main: just triggers the first link in the chain of function-calls
function main() {
  checkInitialState();
}

function restartAll() {
  always.beginElement();
  whenNotActive.beginElement();
  never.beginElement();
}

function checkInitialState() {
  // At first everything should be starting at 1s
  is(always.getStartTime(), 1);
  is(whenNotActive.getStartTime(), 1);
  is(never.getStartTime(), 1);

  // Now try to restart everything early, should be allowed by all
  var restartTime = svg.getCurrentTime();
  restartAll();
  setTimeout('doCheckInitialState(' + restartTime + ')',0);
}
function doCheckInitialState(restartTime) {
  is(always.getStartTime(), restartTime);
  is(whenNotActive.getStartTime(), restartTime);
  is(never.getStartTime(), restartTime);

  // Now skip to half-way
  var newTime = restartTime + 0.5 * always.getSimpleDuration();
  svg.setCurrentTime(newTime);
  setTimeout('checkHalfwayState()', 0);
}

function checkHalfwayState() {
  // Try to restart half-way through--only 'always' should succeed
  var restartTime = svg.getCurrentTime();
  restartAll();
  setTimeout('doCheckHalfwayState(' + restartTime + ')',0);
}
function doCheckHalfwayState(restartTime) {
  is(always.getStartTime(), restartTime);
  isnot(whenNotActive.getStartTime(), restartTime);
  isnot(never.getStartTime(), restartTime);

  // Now skip to the end 
  var newTime = always.getStartTime() + always.getSimpleDuration() + 1;
  svg.setCurrentTime(newTime);
  setTimeout('checkEndingState()', 0);
}

function checkEndingState() {
  // Try to restart after all animations have finished--'always' and
  // 'whenNotActive' should succeed
  var restartTime = svg.getCurrentTime();
  restartAll();
  setTimeout('doCheckEndingState(' + restartTime + ')',0);
}
function doCheckEndingState(restartTime) {
  is(always.getStartTime(), restartTime);
  is(whenNotActive.getStartTime(), restartTime);
  isnot(never.getStartTime(), restartTime);
  SimpleTest.finish();
}

window.addEventListener("load", main, false);
]]>
</script>
</pre>
</body>
</html>
